requirements.txt
MarkupSafe==2.1.5
matplotlib==3.9.1
numpy==1.26.4
pandas==2.2.2
pillow==10.4.0
# transfer-learning/cnn Specific
torch==2.2.2
torchvision==0.17.2
